requirements.txt
matplotlib==3.0.2
sympy==1.6
Flask==1.1.1
line-bot-sdk==1.8.0
